Integrate LNP into core-readout structure
